# openapi.yaml
...
openapi: 3.1.0
info:
  title: VintaSend Managed Templates API
  version: 1.0.0
  description: 'HTTP contract for managing VintaSend notification templates: their
    versions, their status lifecycle, and previewing a version before it is published.'
paths:
  /api/v1/capabilities:
    get:
      operationId: vintasend_templates_management_api_templates_manager_api_get_capabilities
      summary: Get Capabilities
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_dict_str__bool__'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'Which filters the configured template backend can honour.


        Note there are no ``orderBy.*`` keys: the template-manager seam takes no ordering

        argument, so the list endpoint offers no ordering to negotiate. See ``capabilities.py``.'
      tags:
      - system
      security:
      - ApiKeyAuth: []
  /api/v1/templates:
    get:
      operationId: vintasend_templates_management_api_templates_manager_api_list_templates
      summary: List Templates
      parameters:
      - in: query
        name: page
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
        required: false
      - in: query
        name: pageSize
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Pagesize
          type: integer
        required: false
      - in: query
        name: key
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
        required: false
      - in: query
        name: name
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        required: false
      - in: query
        name: description
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        required: false
      - in: query
        name: templateManagedBackend
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Templatemanagedbackend
        required: false
      - in: query
        name: version
        schema:
          anyOf:
          - minimum: 1
            type: integer
          - type: 'null'
          title: Version
        required: false
      - in: query
        name: status
        schema:
          anyOf:
          - items:
              enum:
              - draft
              - active
              - inactive
              - archived
              type: string
            type: array
          - type: 'null'
          title: Status
        required: false
      - in: query
        name: createdAtFrom
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Createdatfrom
        required: false
      - in: query
        name: createdAtTo
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Createdatto
        required: false
      - in: query
        name: updatedAtFrom
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updatedatfrom
        required: false
      - in: query
        name: updatedAtTo
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updatedatto
        required: false
      - in: query
        name: includesAllTags
        schema:
          anyOf:
          - items:
              type: string
            maxItems: 50
            type: array
          - type: 'null'
          title: Includesalltags
        required: false
      - in: query
        name: includesAnyOfTags
        schema:
          anyOf:
          - items:
              type: string
            maxItems: 50
            type: array
          - type: 'null'
          title: Includesanyoftags
        required: false
      - in: query
        name: mostRecentActiveVersion
        schema:
          default: true
          title: Mostrecentactiveversion
          type: boolean
        required: false
      - in: query
        name: isAbstract
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Isabstract
        required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_ManagedTemplateOut_'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'List templates matching the filters -- one row per key by default.


        A row in the store is a *version*, so an unfiltered read of the seam returns
        a key once

        per version it has ever had. ``mostRecentActiveVersion`` defaults to ``true``
        and asks

        the backend for the current version of each key instead: the highest-numbered
        ``active``

        or ``draft`` one. The collapsing happens in the store, not here, so a page
        is still a

        page of what the backend counted.


        Send ``mostRecentActiveVersion=false`` for the raw listing, every version
        included. A

        backend that cannot answer the filter has it dropped like any other unsupported
        one, and

        then returns every version too.'
      tags:
      - templates
      security:
      - ApiKeyAuth: []
    post:
      operationId: vintasend_templates_management_api_templates_manager_api_create_template
      summary: Create Template
      parameters: []
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_ManagedTemplateOut_'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'Create a template''s first version.


        Whether re-using an existing key is an error is the backend''s call, not this
        API''s --

        the seam does not define it, and a backend that treats a repeat ``create_template``
        as

        a new version is behaving legitimately. Use ``POST /templates/{key}/versions``
        when you

        mean "next version of this key".'
      tags:
      - templates
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateBody'
        required: true
      security:
      - ApiKeyAuth: []
  /api/v1/templates/{key}/versions:
    get:
      operationId: vintasend_templates_management_api_templates_manager_api_list_template_versions
      summary: List Template Versions
      parameters:
      - in: path
        name: key
        schema:
          title: Key
          type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse_ManagedTemplateOut_'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'Every version of one template, newest version first.


        The service resolves this by filtering on the key, which matches nothing for
        a key that

        does not exist rather than raising -- so an empty result is a missing key,
        and is

        reported as the 404 the contract documents rather than an empty list. A key
        that exists

        always has at least one version.'
      tags:
      - templates
      security:
      - ApiKeyAuth: []
    post:
      operationId: vintasend_templates_management_api_templates_manager_api_create_template_version
      summary: Create Template Version
      parameters:
      - in: path
        name: key
        schema:
          title: Key
          type: string
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_ManagedTemplateOut_'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'Create a new version of an existing template, copied forward from
        its latest one.


        Templates are versioned rather than edited in place, which is why this is
        a POST that

        creates a resource and not a PATCH that mutates one: an already-published
        version is

        never modified. Fields left unset are carried over from the latest version.'
      tags:
      - templates
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVersionBody'
        required: true
      security:
      - ApiKeyAuth: []
  /api/v1/templates/{key}/versions/{version}:
    get:
      operationId: vintasend_templates_management_api_templates_manager_api_get_template_version
      summary: Get Template Version
      parameters:
      - in: path
        name: key
        schema:
          title: Key
          type: string
        required: true
      - in: path
        name: version
        schema:
          title: Version
          type: integer
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_ManagedTemplateOut_'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      tags:
      - templates
      security:
      - ApiKeyAuth: []
    delete:
      operationId: vintasend_templates_management_api_templates_manager_api_delete_template_version
      summary: Delete Template Version
      parameters:
      - in: path
        name: key
        schema:
          title: Key
          type: string
        required: true
      - in: path
        name: version
        schema:
          title: Version
          type: integer
        required: true
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      tags:
      - templates
      security:
      - ApiKeyAuth: []
  /api/v1/templates/{key}/composition:
    get:
      operationId: vintasend_templates_management_api_templates_manager_api_get_template_composition
      summary: Get Template Composition
      parameters:
      - in: path
        name: key
        schema:
          title: Key
          type: string
        required: true
      - in: query
        name: version
        schema:
          anyOf:
          - minimum: 1
            type: integer
          - type: 'null'
          title: Version
        required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_TemplateCompositionOut_'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'One version assembled the way the template engine will receive
        it.


        A managed template can build on another: ``{% managed_extends "base" %}``
        to fill a

        base''s ``{% managed_children %}`` hole and override its ``{% managed_block
        %}`` regions,

        ``{% managed_include "footer" %}`` to splice a fragment in. All of it is resolved
        before

        the engine runs, so the stored ``bodyTemplate`` is only half the template
        -- this is what

        actually renders.


        Nothing here is rendered against a context: engine syntax survives untouched.
        Use

        ``POST /templates/{key}/preview`` for the rendered result.


        Also reports what this version directly references, and whether it is abstract
        --

        recomputed from the source rather than read from the stored flag, so it is
        the authority

        behind the ``isAbstract`` on every template payload.


        A template that cannot be assembled -- a base that does not exist, a chain
        that loops, a

        malformed tag -- comes back as a 409 ``TEMPLATE_COMPOSITION_ERROR`` naming
        the chain that

        broke, for the same reason a template that will not render is a 409 rather
        than a 500.'
      tags:
      - templates
      security:
      - ApiKeyAuth: []
  /api/v1/templates/{key}/status-history:
    get:
      operationId: vintasend_templates_management_api_templates_manager_api_get_status_history
      summary: Get Status History
      parameters:
      - in: path
        name: key
        schema:
          title: Key
          type: string
        required: true
      - in: query
        name: version
        schema:
          anyOf:
          - minimum: 1
            type: integer
          - type: 'null'
          title: Version
        required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse_TemplateStatusHistoryOut_'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'The status audit trail, most recent change first.


        Omitting ``version`` asks for the whole key''s history, which backends that
        keep it that

        way will return. Unlike everywhere else in this API, ``version=None`` here
        does not mean

        "the latest version".'
      tags:
      - statuses
      security:
      - ApiKeyAuth: []
...
  /api/v1/templates/{key}/deactivate:
    post:
      operationId: vintasend_templates_management_api_templates_manager_api_deactivate_template
      summary: Deactivate Template
      parameters:
      - in: path
        name: key
        schema:
          title: Key
          type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_ManagedTemplateOut_'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: Retire one version without archiving it, so it can be activated
        again later.
      tags:
      - statuses
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/StatusChangeBody'
              default:
                version: null
                changedBy: null
        required: false
      security:
      - ApiKeyAuth: []
...
components:
  schemas:
    DataResponse_dict_str__bool__:
      properties:
        data:
          additionalProperties:
            type: boolean
          title: Data
          type: object
      required:
      - data
      title: DataResponse[dict[str, bool]]
      type: object
...
    ApiErrorResponse:
      description: Error envelope returned with every non-2xx response.
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorBody'
      required:
      - error
      title: ApiErrorResponse
      type: object
...
    PaginatedResponse_ManagedTemplateOut_:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ManagedTemplateOut'
          title: Data
          type: array
        page:
          title: Page
          type: integer
        pageSize:
          title: Pagesize
          type: integer
        hasMore:
          title: Hasmore
          type: boolean
      required:
      - data
      - page
      - pageSize
      - hasMore
      title: PaginatedResponse[ManagedTemplateOut]
      type: object
    DataResponse_ManagedTemplateOut_:
      properties:
        data:
          $ref: '#/components/schemas/ManagedTemplateOut'
      required:
      - data
      title: DataResponse[ManagedTemplateOut]
      type: object
...
    ListResponse_ManagedTemplateOut_:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ManagedTemplateOut'
          title: Data
          type: array
      required:
      - data
      title: ListResponse[ManagedTemplateOut]
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
servers: []